refactor: 체스 게임 패키지 분리
