+++ src/main/java/chess/domain/game/state/finished/FinishedGame.java
@@ -1,5 +1,7 @@
-package chess.domain.game.state;
+package chess.domain.game.state.finished;
 
+import chess.domain.game.state.ChessGame;
+import chess.domain.game.state.started.StartedGame;
 import chess.domain.piece.Camp;
 import chess.domain.position.ChessBoard;
 import chess.domain.position.Position;
@@ -8,7 +10,7 @@ public abstract class FinishedGame extends StartedGame {
 
     private static final String GAME_ALREADY_END = "이미 게임이 종료 되었습니다.";
 
-    FinishedGame(ChessBoard chessBoard, Camp turnCamp) {
+    protected FinishedGame(ChessBoard chessBoard, Camp turnCamp) {
         super(chessBoard, turnCamp);
     }
